requirements.txt
# -- Packages that should have been installed --
# torch
# torchvision
# -- Packages that will be installed --
pytorchvideo
ffmpeg-python
augly
h5py
pandas
tqdm
opencv-python
scikit-learn
einops
timm
typer
tensorboard
seaborn
imageio
faiss-cpu
fastapi
uvicorn[standard]
python-multipart
# Optional: pip install open-clip-torch  (for CLIP embedding backend)
# Optional: pip install vpdq             (for perceptual hash pre-filter)
# Optional: pip install qdrant-client    (for Qdrant vector database backend)